creacion de entidades, controladores, servicios y repositorios de productos y categorias funcionalidad de crear y eliminar productos y ver productos de vendedores

+++ src/main/java/hn/unah/ingenieria/pu_market/repository/usuarioRepositorio.java
@@ -9,6 +9,6 @@ import hn.unah.ingenieria.pu_market.entity.Usuario;
 
 
 @Repository
-public interface usuarioRepositorio extends JpaRepository<Usuario, Long>{
+public interface usuarioRepositorio extends JpaRepository<Usuario, Integer>{
     Optional<Usuario> findByCorreoInstitucional(String correoInstitucional);
 }
