+++ src/main/java/hn/unah/ingenieria/pu_market/repository/verificacionesRepositorio.java
@@ -9,7 +9,7 @@ import hn.unah.ingenieria.pu_market.entity.Usuario;
 import hn.unah.ingenieria.pu_market.entity.Verificacion;
 
 @Repository
-public interface verificacionesRepositorio extends JpaRepository<Verificacion, Long> {
+public interface verificacionesRepositorio extends JpaRepository<Verificacion, Integer> {
     Optional<Verificacion> findByToken(String token);
     Optional<Verificacion> findByUsuario(Usuario usuario);
